pesquisador_page

+++ src/pesquisador/List.jsx
@@ -2,24 +2,19 @@ import { useState, useRef } from "react"
 import styles from "./List.module.css"
 import pesquisadores from './pesquisadores.json'
 import pesquisadoresObras from './pesquisador_obras.json'
-// const pesquisadores = require('pesquisadores.json')
 
-// import { useState, useRef } from "react"
-
-const List = ({ result, setQueried, queried }) => {
-    // console.log(pesquisadores)
-    // const pesquisadores = require('pesquisadores.json')
+const List = ({ result, institutos, setQueried, queried }) => {
     const [isModalOpen, setIsModalOpen] = useState(false)
     const [isUpdateModalOpen, setIsUpdateModalOpen] = useState(false)
-    const [toUpdateNome, setToUpdateNome] = useState("")
-    const [toUpdateAcr, setToUpdateAcr] = useState("")
-    const [toUpdateNewNome, setToUpdateNewNome] = useState("")
     const [toAddId, setToAddId] = useState("")
     const [toAddInstituto, setToAddInstituto] = useState("")
-    const [toAdd, setToAdd] = useState("")
-  
+    const [toUpdateNome, setToUpdateNome] = useState("")
+    const [toUpdateEmail, setToUpdateEmail] = useState("")
+    const [toUpdateInstituto, setToUpdateInstituto] = useState("")  
+
     const updateNameRef = useRef([])
-    const updateAcrRef = useRef([])
+    const updateEmailRef = useRef([])
+    const updateInstitutoRef = useRef([])
   
     const changeQueried = () => {
       setQueried(!queried)
@@ -42,47 +37,37 @@ const List = ({ result, setQueried, queried }) => {
       const instituto = await response.json()
 
       if(instituto.length == 0) return
-
-      console.log(pesquisadoresObras[0]['artigos'])
-
-      console.log(pesquisador)
-
+      
       pesquisador['artigos'].forEach(artigo => {
-
-        console.log(artigo['titulo'])
-        console.log(artigo['ano'])
-        console.log(artigo['status'])
+        
         fetch(`http://localhost:8080/api/v1/obras/postObra?titulo=${artigo['titulo']}&ano=${artigo['ano']}&status[${artigo['status']}]`,       
         {
           method: "POST",
         })
-
-      });
-
-      // const data = await response.json()
-      // console.log(data)
-      
-      // .then(response => {
         
-      //   if(response.status == 200){
-      //     console.log(response.json())
-      //   }
+      });
       
-      
-      // })
-
-      // fetch(
-      //   `http://localhost:8080/api/v1/pesquisador/post?id=${toAddId}&instituto=${toAddInstituto}`,
-      //   {
-      //     method: "POST",
-      //   }
-      // ).then(() => changeQueried())
-      // isModalOpen = !isModalOpen
+      fetch(`http://localhost:8080/api/v1/pesquisadores/post?idXml=${pesquisador['idXml']}&nome=${pesquisador['nome']}&idInstituto=${instituto[0]['id']}`,       
+      {
+        method: "POST",
+      })
+      .then(() => changeQueried())
     }
   
-    const handleUpdateElementClick = () => {
+    const handleUpdateElementClick = async () => {
+
+      const response = await fetch(`http://localhost:8080/api/v1/institutos/institutoNome?nome=${toUpdateInstituto}`)
+
+      const instituto = await response.json()
+
+      // console.log
+
+      if(instituto.length == 0) return
+
+      console.log(`http://localhost:8080/api/v1/pesquisadores/update?nome=${toUpdateNome}&instituto=${instituto[0]['id']}&email=${toUpdateEmail}`)
+
       fetch(
-        `http://localhost:8080/api/v1/institutos/update?newNome=${toUpdateNewNome}&nome=${toUpdateNome}&acronimo=${toUpdateAcr}`,
+        `http://localhost:8080/api/v1/pesquisadores/update?nome=${toUpdateNome}&instituto=${instituto[0]['id']}&email=${toUpdateEmail}`,
         {
           method: "PUT",
         }
@@ -92,18 +77,32 @@ const List = ({ result, setQueried, queried }) => {
     }
   
     const handleDeleteElement = (e) => {
-      fetch(`http://localhost:8080/api/v1/institutos/delete?nome=${e.innerText}`, {
+
+      const pesquisador = pesquisadoresObras.find(p => p.nome == e.innerText)
+
+      if (pesquisador === undefined) return
+
+      pesquisador['artigos'].forEach(artigo => {
+
+        fetch(`http://localhost:8080/api/v1/obras/deleteObra?titulo=${artigo['titulo']}`,
+        {
+          method: "DELETE",
+        })
+
+      })
+
+      fetch(`http://localhost:8080/api/v1/pesquisadores/delete?idXml=${pesquisador['idXml']}`,
+      {
         method: "DELETE",
       })
-      // .then(response => console.log(response))
       .then(() => changeQueried())
+
     }
   
     const handleUpdateModalOpen = (i) => {
-      // console.log(updateNameRef)
       setToUpdateNome(updateNameRef.current[i].innerText)
-      setToUpdateNewNome(updateNameRef.current[i].innerText)
-      setToUpdateAcr(updateAcrRef.current[i].innerText)
+      updateEmailRef.current[i].innerText == 'NA' ? setToUpdateEmail('') : setToUpdateEmail(updateEmailRef.current[i].innerText)
+      setToUpdateInstituto(updateInstitutoRef.current[i].innerText)
       setIsUpdateModalOpen(true)
     }
   
@@ -159,23 +158,23 @@ const List = ({ result, setQueried, queried }) => {
               </button>
               <div className={styles.interiorAddModalContainer}>
                 <span>
-                  <label htmlFor="name">Nome</label>
+                  <label htmlFor="email">Email</label>
                   <input
                     type="text"
                     className={styles.newName}
-                    placeholder="Novo nome..."
-                    value={toUpdateNewNome}
-                    onChange={(e) => setToUpdateNewNome(e.target.value.trim())}
+                    placeholder="Novo email..."
+                    value={toUpdateEmail}
+                    onChange={(e) => setToUpdateEmail(e.target.value.trim())}
                   />
                 </span>
                 <span>
-                  <label htmlFor="abv">Abreviação</label>
+                  <label htmlFor="inst">Instituto</label>
                   <input
                     type="text"
                     className={styles.newAbv}
-                    placeholder="Novo acrônimo..."
-                    value={toUpdateAcr}
-                    onChange={(e) => setToUpdateAcr(e.target.value.trim())}
+                    placeholder="Novo Instituto..."
+                    value={toUpdateInstituto}
+                    onChange={(e) => setToUpdateInstituto(e.target.value.trim())}
                   />
                 </span>
                 <span>
@@ -200,23 +199,33 @@ const List = ({ result, setQueried, queried }) => {
           <div className={styles.list_header}>
             <span className={styles.header_name}>Nome</span>
             <span className={styles.header_abv}>E-mail</span>
-            <span className={styles.header_actionBtns}>Institutos</span>
+            <span className={styles.header_actionBtns}>Instituto</span>
+            <span className={styles.header_actionBtns}>Acoes</span>
           </div>
           {result.length > 0 &&
             result.map((e, i) => {
+              
+              const instituto = institutos.find(inst => inst['id'] == e['idinstituto'])
+              // console.log(e)
               return (
                 <div key={`list${i}`} className={styles.list_line}>
                   <span
-                    className={styles.name}
+                    // className={styles.pesquisador_field}
                     ref={(el) => (updateNameRef.current[i] = el)}
                   >
-                    {e.nome}
+                    {e.nome || 'NA'}
                   </span>
                   <span
-                    className={styles.abv}
-                    ref={(el) => (updateAcrRef.current[i] = el)}
+                    // className={styles.pesquisador_field}
+                    ref={(el) => (updateEmailRef.current[i] = el)}
                   >
-                    {e.acronimo}
+                    {e.email || 'NA'}
+                  </span>
+                  <span
+                    // className={styles.pesquisador_field}
+                    ref={(el) => (updateInstitutoRef.current[i] = el)}
+                  >
+                    {instituto.nome || 'NA'}
                   </span>
                   <span className={styles.actionBtns}>
                     <button
